requirements/test.txt
-r prod.txt
tox
coverage
# fixing version to avoid bug with Windows subst paths
pytest==5.0.1
pytest-cov
pytest-flask==0.15.1
# pytest-docker
# docker-compose
pyyaml==4.2b4
psycopg2==2.8.3
sqlalchemy_utils
